update with new SOMA services
