+++ activity_checking/config/region_names.yaml
@@ -1,40 +1,51 @@
 {
+    "8": "Robot lab"
+    "9": "Robot lab"
+    "10": "Robot lab"
+    "11": "Robot lab"
+    "1": "Student's Area"
+    "2": "Student's Area"
+    "3": "Student's Area"
+    "4": "Student's Area"
+    "5": "Student's Area"
+    "6": "Atrium Area"
+    "7": "Atrium Area"
     # "21": "PhD Workspace",
     # "22": "PhD Workspace",
     # "23": "PostDoc Workspace"
-    "110": "Kitchen",
-    "107": "Hospitality",
-    "108": "Hospitality",
-    "109": "Hospitality",
-    "129": "Printing Area",
-    "111": "Kitchen",
-    "71": "Corperative Services",
-    "72": "Corperative Services",
-    "114": "Corperative Services",
-    "116": "Corperative Services",
-    "126": "Corperative Services",
-    "124": "Corperative Services",
-    "125": "Corperative Services",
-    "119": "Corperative Services",
-    "118": "Corperative Services",
-    "62": "Corperative Services",
-    "73": "Corperative Services",
-    "74": "Corperative Services",
-    "130": "Printing Area",
-    "127": "Eating Area",
-    "131": "Eating Area",
-    "132": "Eating Area",
-    "134": "Eating Area",
-    "128": "Kitchen",
-    "133": "Kitchen",
-    "137": "Supportive Services",
-    "138": "Supportive Services",
-    "140": "Supportive Services",
-    "136": "Supportive Services",
-    "86": "Supportive Services",
-    "87": "Supportive Services",
-    "135": "Supportive Services",
-    "146": "Supportive Services",
-    "144": "Supportive Services",
-    "142": "Supportive Services",
+    # "110": "Kitchen",
+    # "107": "Hospitality",
+    # "108": "Hospitality",
+    # "109": "Hospitality",
+    # "129": "Printing Area",
+    # "111": "Kitchen",
+    # "71": "Corperative Services",
+    # "72": "Corperative Services",
+    # "114": "Corperative Services",
+    # "116": "Corperative Services",
+    # "126": "Corperative Services",
+    # "124": "Corperative Services",
+    # "125": "Corperative Services",
+    # "119": "Corperative Services",
+    # "118": "Corperative Services",
+    # "62": "Corperative Services",
+    # "73": "Corperative Services",
+    # "74": "Corperative Services",
+    # "130": "Printing Area",
+    # "127": "Eating Area",
+    # "131": "Eating Area",
+    # "132": "Eating Area",
+    # "134": "Eating Area",
+    # "128": "Kitchen",
+    # "133": "Kitchen",
+    # "137": "Supportive Services",
+    # "138": "Supportive Services",
+    # "140": "Supportive Services",
+    # "136": "Supportive Services",
+    # "86": "Supportive Services",
+    # "87": "Supportive Services",
+    # "135": "Supportive Services",
+    # "146": "Supportive Services",
+    # "144": "Supportive Services",
+    # "142": "Supportive Services",
 }
